Refactor collaboration room code
