+++ frontend/src/components/CollaborationRoom/ConfirmationBox.tsx
