+++ frontend/src/pages/Room.tsx
@@ -1,7 +1,8 @@
 import { useState } from 'react'
 import { useNavigate } from 'react-router'
-import CodeEditor from '../components/CodeEditor/CodeEditor'
-import ConfirmationBox from '../components/CodeEditor/ConfirmationBox.tsx'
+import CodeEditor from '../components/CollaborationRoom/CodeEditor'
+import ConfirmationBox from '../components/CollaborationRoom/ConfirmationBox.tsx'
+import QuestionDescription from '../components/CollaborationRoom/QuestionDescription.tsx'
 
 export const Room = () => {
     const [showConfirmation, setShowConfirmation] = useState(false)
@@ -17,7 +18,7 @@ export const Room = () => {
             {showConfirmation && (
                 <ConfirmationBox onClose={() => setShowConfirmation(false)} onExit={handleExit} />
             )}
-            <nav className='nav' style={{ padding: ' 8px 20px' }}>
+            <nav className='nav' style={{ padding: '8px 20px' }}>
                 <h2
                     style={{
                         margin: '0 0 0 2rem',
@@ -36,24 +37,8 @@ export const Room = () => {
                 </button>
             </nav>
             <div className='split-container' style={{ padding: '20px' }}>
-                <div
-                    className='pane'
-                    style={{
-                        backgroundColor: '#303030',
-                        marginRight: '10px',
-                        padding: '1.5rem 2rem',
-                    }}
-                >
-                    <h2 style={{ margin: '0', fontWeight: 'normal', fontSize: '1.75rem' }}>
-                        Question Title
-                    </h2>
-                    <h2
-                        className={`complexity-color-Easy`}
-                        style={{ margin: '0', fontSize: '1.25rem' }}
-                    >
-                        Easy
-                    </h2>
-                    <div style={{ marginTop: '1.5rem' }}>Question description</div>
+                <div className='pane' style={{ backgroundColor: '#303030', marginRight: '10px' }}>
+                    <QuestionDescription />
                 </div>
                 <div className='pane' style={{ backgroundColor: '#303030', marginLeft: '10px' }}>
                     <CodeEditor />
